feat: improved image upload and GridFS integration

+++ src/services/Post.services.ts
@@ -1,23 +1,29 @@
+
+import { ObjectId } from "mongodb";
 import { CreatePostDTO, UpdatePostDTO } from "../dtos/PostDTO";
 import { ThrowError } from "../errors/ThrowError";
 import Filter from "../interfaces/Filter";
 import IService from "../interfaces/IService";
 import IPost from "../models/Post";
 import PostRepository from "../repositories/Post.repository";
+import { PictureStorageRepository } from "../repositories/PictureStorage.repository";
 
 const postRepository = new PostRepository();
 export class PostService implements IService<CreatePostDTO, UpdatePostDTO, IPost> {
-    getPostsByAccount(accountId: string) {
+
+    async getPostsByAccount(accountId: string) {
         try {
-            return postRepository.getPostsByAccount(accountId);
+            return await postRepository.getPostsByAccount(accountId);
         } catch (error) {
             if (error instanceof ThrowError) throw error;
             throw ThrowError.internal("Não foi possível buscar os posts.");
         }
     }
-    async getPostsWithAuthor(filter: Filter) {
+    async getPostsWithAuthor(filter: Filter): Promise<IPost[]> {
         try {
-            return await postRepository.getPostsWithAuthor(filter);
+            const posts = await postRepository.getPostsWithAuthor(filter);
+            const post = posts.map((post) => ({ ...post, id: post?._id.toString() } as IPost));
+            return post;
         } catch (error) {
             if (error instanceof ThrowError) throw error;
             throw ThrowError.internal("Não foi possível buscar os posts.");
@@ -25,22 +31,22 @@ export class PostService implements IService<CreatePostDTO, UpdatePostDTO, IPost
     }
     async toggleLike(postId: string, accountId: string): Promise<IPost | null> {
         try {
+            if (!postId || !accountId) return null;
             const post = await postRepository.getById(postId);
             if (!post) return null;
 
             const alreadyLiked = post.likes.some((id) => id.equals(accountId));
-
+            
             return alreadyLiked
                 ? await postRepository.removeLike(postId, accountId)
-                : await postRepository.addLike(postId, accountId);
-        } catch (error) {
+                : await postRepository.addLike(postId, accountId);;
+        } catch (error: any) {
             if (error instanceof ThrowError) throw error;
             throw ThrowError.internal("Não foi possível curtir o post.");
         }
     }
     updateComment(accountId: string, updateData: UpdatePostDTO): Promise<IPost | null> {
         try {
-
             return postRepository.update(accountId, updateData);
         } catch (error) {
             if (error instanceof ThrowError) throw error;
@@ -68,8 +74,17 @@ export class PostService implements IService<CreatePostDTO, UpdatePostDTO, IPost
         }
     }
 
-    async create(data: CreatePostDTO): Promise<IPost> {
+    async create(data: CreatePostDTO, files?: Express.Multer.File[]): Promise<IPost> {
+        const images: ObjectId[] = [];
         try {
+            if (files && files.length > 0) {
+                for (const file of files) {
+                    const id = await PictureStorageRepository.uploadImage(file);
+                    if(!id) continue;
+                    images.push(id);
+                }
+            }
+            data.image = images;
             return await postRepository.create(data);
         } catch (error: any) {
             if (error instanceof ThrowError) throw error;
